testing: rename package directory to axontest
The testing/ directory declared package axontest but had a mismatched
import path, forcing every caller to write an explicit alias. Renaming
to axontest/ aligns the directory with the package name (same pattern
as net/http/httptest) and lets importers drop the alias entirely.

Updates go.work, examples/go.mod, and all eight example import sites.
