+++ examples/01-minimal/main.go
@@ -17,8 +17,8 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/axonframework/axon/axontest"
 	"github.com/axonframework/axon/kernel"
-	axontest "github.com/axonframework/axon/testing"
 )
 
 func main() {
