+++ examples/go.mod
@@ -10,7 +10,7 @@ require (
 	github.com/axonframework/axon/plan v0.0.0
 	github.com/axonframework/axon/providers/anthropic v0.0.0
 	github.com/axonframework/axon/providers/openai v0.0.0
-	github.com/axonframework/axon/testing v0.0.0
+	github.com/axonframework/axon/axontest v0.0.0
 	github.com/axonframework/axon/workflow v0.0.0
 	github.com/openai/openai-go/v3 v3.32.0
 )
@@ -30,6 +30,6 @@ replace (
 	github.com/axonframework/axon/plan => ../plan
 	github.com/axonframework/axon/providers/anthropic => ../providers/anthropic
 	github.com/axonframework/axon/providers/openai => ../providers/openai
-	github.com/axonframework/axon/testing => ../testing
+	github.com/axonframework/axon/axontest => ../axontest
 	github.com/axonframework/axon/workflow => ../workflow
 )
